moved GossipedBlock to seperate file instead of events.go
