+++ proto/p2p/gossiped_block.proto
